Variable type change from uint to size_t when calling functions in vfs
git-svn-id: http://svn.wildfiregames.com/public/ps/trunk@4230 3db68df2-c116-0410-a063-a993310a9797

+++ source/ps/ConfigDB.cpp
@@ -244,7 +244,7 @@ bool CConfigDB::Reload(EConfigNamespace ns)
 	parser.InputTaskType("CommentOrBlank", "_[;[$rest]]");
 
 	FileIOBuf buffer;
-	uint buflen;
+	size_t buflen;
 	File f;
 	LibError ret;
 	if (m_UseVFS[ns])
